requirements.txt
flask
python-dotenv
PyPDF2
gunicorn
requests
